refactor: 각 파일 적합한 폴더로 이동 및 경로 수정 - project-detail 타입을 src/types로 이동 - 컴포넌트 및 훅의 import 경로 수정
