+++ src/components/project-detail/index.ts
@@ -1,4 +1,3 @@
 export { default as ApplyModal } from "./ApplyModal";
 export { default as SuccessToast } from "./SuccessToast";
-export * from "./types";
 export * from "./constants";
