+++ src/types/project-detail.ts
